fix(tooltip-web): arrow placement and styling
